Added information about the caching for deep health checks
Rx uses an in-memory cache for deep health checks by default.
This can cause some problems for example when you are writing tests for
the deep health check endpoint that uses the library.

The problem can be solved by disabling the cache, however this was
previously not discussed in the docs.