allow configuration of health check paths
the default healthcheck paths (/liveness, /readiness, and /deep)
can be now be configured by passing [type]_path keys to the options
hash during initialization. to change the liveness path, for example:

Rx::Middleware.new(options: { liveness_path: "/custom" })

this change is opt-in; by default these options are not expected and
will default to the original paths.